scss for the lobby menu
