+++ frontend/src/App.scss
@@ -1,5 +1,6 @@
 body {
-  background-color: seagreen;
+  background: darkgreen;
+  background: radial-gradient(circle, seagreen 0%, darkgreen 100%);
   font-family: "Fira Sans", sans-serif;
   font-weight: 400;
   color: #ededed;
