+++ frontend/src/App.tsx
@@ -1,7 +1,7 @@
 import './App.scss'
 import {Navigate, Routes, Route} from "react-router-dom";
-import Lobby from './pages/Lobby';
-import Game from "./pages/Game.tsx";
+import Lobby from './pages/Lobby/Lobby.tsx';
+import Game from "./pages/Game/Game.tsx";
 
 export default function App() {
